entries-api/filesystemdirectoryentry-manual.html

<!DOCTYPE html>
<meta charset=utf-8>
<title>Entries API: FileSystemDirectoryEntry manual test</title>
<link rel=help href="https://wicg.github.io/entries-api/#api-directoryentry">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="support.js"></script>

<script>
entry_test((t, entry) => {
  // FileSystemEntry ----------

  assert_idl_attribute(entry, 'isFile', 'FileSystemDirectoryEntry has isFile attribute');
  assert_equals(typeof entry.isFile, 'boolean', 'isFile is boolean');

  assert_idl_attribute(entry, 'isDirectory', 'FileSystemDirectoryEntry has isDirectory attribute');
  assert_equals(typeof entry.isDirectory, 'boolean', 'isDirectory is boolean');

  assert_idl_attribute(entry, 'name', 'FileSystemDirectoryEntry has name attribute');
  assert_equals(typeof entry.name, 'string', 'name is a string');

  assert_idl_attribute(entry, 'fullPath', 'FileSystemDirectoryEntry has fullPath attribute');
  assert_equals(typeof entry.fullPath, 'string', 'fullPath is a string');

  assert_idl_attribute(entry, 'filesystem', 'FileSystemDirectoryEntry has filesystem attribute');

  assert_idl_attribute(entry, 'getParent', 'FileSystemDirectoryEntry has getParent');
  assert_equals(typeof entry.getParent, 'function', 'getParent() is a method');

  // FileSystemDirectoryEntry  ----------

  assert_idl_attribute(entry, 'getFile', 'FileSystemDirectoryEntry has getFile');
  assert_equals(typeof entry.getFile, 'function',
                'getFile() is a method');

  assert_idl_attribute(entry, 'getDirectory', 'FileSystemDirectoryEntry has getDirectory');
  assert_equals(typeof entry.getDirectory, 'function',
                'getDirectory() is a method');

  assert_idl_attribute(entry, 'createReader', 'FileSystemDirectoryEntry has createReader');
  assert_equals(typeof entry.createReader, 'function',
                'createReader() is a method');

  t.done();
}, 'FileSystemDirectoryEntry - interface');

entry_test((t, entry) => {
  assert_false(entry.isFile, 'isFile is false');
  assert_true(entry.isDirectory, 'isDirectory is true');
  assert_equals(entry.name, 'upload', 'expected directory was uploaded');
  assert_equals(entry.fullPath, '/upload', 'directory is child of root directory');
  t.done();
}, 'FileSystemDirectoryEntry - attributes');
</script>
